Use gecko-mapfile to find mozilla commit
The try-prepare-repo script was using the 'FIREFOX_.*' tags to match the
firefox commit in git with the matching commit in mercurial.
Unfortunately those tags are no longer present in the gecko-dev git
repository.

So instead of using the 'FIREFOX_.*' tags we are now using the
gecko-mapfile file which gives us the mapping between git and mercurial
commit hashes. The script 'find-hg-parent-commit' is used to find the
last mozilla commit in tor-browser history, and provide the git and
mercurial hash of this commit.